Pandas and Google LSA Project/Google LSA Automation Project/Final Automation Tool_ Customer to BigQuery/requirements.txt
APScheduler==3.9.1
bigquery==0.0.11
requests==2.27.1
google-api-core==2.8.1
google-api-python-client==1.7.11
google-auth==2.6.6
google-auth-oauthlib==0.5.1
numpy==1.22.4
pandas==1.4.2
gspread==3.7.0
oauth2client==4.1.3
oauthlib==3.2.0
pytz==2020.4
rsa==4.6
tzlocal==4.2